Worked on some of the failing test cases, too. No test is succeeding on a subsequent run anymore. Ported the test cases TestWorkbookEvaluator, TestEvaluationCache and TestCellCacheEntry (but dropped most of my version).
git-svn-id: http://npoi.googlecode.com/svn/trunk@180 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ testcases/main/HSSF/Record/Formula/Eval/TestCircularReferences.cs
@@ -52,7 +52,7 @@ namespace TestCases.SS.Formula.Eval
         private static void ConfirmCycleErrorCode(NPOI.SS.UserModel.CellValue cellValue)
         {
             Assert.IsTrue(cellValue.CellType == NPOI.SS.UserModel.CellType.ERROR);
-            Assert.AreEqual((byte)ErrorEval.CIRCULAR_REF_ERROR.ErrorCode, cellValue.ErrorValue);
+            Assert.AreEqual(ErrorEval.CIRCULAR_REF_ERROR.ErrorCode, cellValue.ErrorValue);
         }
